pyproject.toml
[project]
name = "git-sync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "y-sweet-sdk>=0.6.3",
    "requests",
    "pycrdt",
    "GitPython",
    "flask",
    "cryptography",
    "python-dateutil",
    "pyjwt",
    "svix",
    "starlette",
    "uvicorn",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "pre-commit",
]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
pythonpath = ["."]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
